Update 2014-03-28-My-Fifth-Post.md
Discovered to my horror that the GitHub template attempts to display code as formatted content!  Removed the code until I can figure out whether it's possible to include code in a formatted post, and included a link to the project_code directory for those hard-core enough to want to see the code.